+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\TimeContinuum\Format\Custom`
+
 ## 4.1.1 - 2025-04-21
 
 ### Fixed